+++ printfactory/__init__.py
@@ -4,7 +4,12 @@ __version__ = version('printfactory')
 
 
 from .printer import *
+from .print_tool import *
+from .print_tools.adobe import *
 
 __all__ = [
-    'Printer'
+    'Printer',
+    'PrintTool',
+    'AdobeReader',
+    'AdobeAcrobat',
 ]