+++ CHANGELOG.md
@@ -7,13 +7,20 @@ and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
 
-### [0.1.0] (2020-12-18)
+### [0.1.0] (2020-02-10)
 
 #### Added
 
 - add basic app skeleton example with tests and docs
-- add function to get a list of installed printers
-- add printer class for sending a file to a printer
+- add _Printer_ class to set up a printer for a PrintTool
+- add a function to get a list of installed printers on your system 
+- add generic _PrintTool_ class for sending files to a printer
+- add _AdobeReader_ class for sending files to a _Printer_
+- add _AdobeAcrobat_ class for sending files to a _Printer_
+
+#### Known issues
+
+- The _AdobeAcrobat_ print tool implementation is limited to only send files to the defaults system printer  
 
 ## [Releases]